Added components folder. Moved AddTodoForm and TodoList components to <components> folder. Changed components extension to jsx
